Bug fix in line 159 pybrain/supervised/trainers/backprop.py: float converted to int

+++ pybrain/supervised/trainers/backprop.py
@@ -156,7 +156,7 @@ class BackpropTrainer(Trainer):
         if verbose:
             print('Average error:', avgErr)
             print(('Max error:', max(ponderatedErrors), 'Median error:',
-                   sorted(ponderatedErrors)[len(errors) / 2]))
+                   sorted(ponderatedErrors)[int(len(errors) / 2)]))
         return avgErr
 
     def testOnClassData(self, dataset=None, verbose=False,
